fix dataset size

+++ libethash/ethash.cpp
@@ -22,7 +22,7 @@ namespace ethash
 constexpr static int light_cache_init_size = 1 << 24;
 constexpr static int light_cache_growth = 1 << 20;
 constexpr static int light_cache_rounds = 3;
-constexpr static int full_dataset_init_size = 1 << 32;
+constexpr static unsigned long long full_dataset_init_size = 1ULL << 32;
 constexpr static int full_dataset_growth = 1 << 25;
 constexpr static int full_dataset_item_parents = 512;
